🚀 [Feat]: shared api 타입 feature 계층으로 이동
